Create array, constructor, and calling function
In the function addBookToLibrary I struggled. I kept calling the this keyword for the arguments, but it naturally was referencing the global object as it's a function invocation. Upun research I came across object.create() from Ben Nadel's article who was trying to solve the same problem. It worked but it didn't use the 'new' keyword. Upon more research I realised that I should treat the parameters as if they were arguments when I call upon a function. The exact same line I used to create one object, I should use in the function and substitute actual information with the parameters.