Try to fix THE BUG
There's a bug that makes the libraryBooksDisplay function run multiple times upon clicking the button. To fix it I removed the second for loop. This didn't help but the loop was unnecessary in the first place. I then did a debugging session to see what was going on and I found that: After all of the cellContents are added to the cells and the cell t appended to the row, the code goes back to the myLibrary for-loop to see if there's any other book to add. In the first button click, It finds just one book and so it adds that book again, but doesn't display it. So there's now two rows but the second one is hidden. It then goes all the way down to add tbody to table and so on. Then When I create a new book, it shows that previous second row plus the new book.
So I thought the logic I should create should be: After you add the last book to the library, don't go back again. But that's dumb, as it won't show the entire database of books. So instead I added a while loop that will remove the first instance of the duplicated books in the table. That helped the table display, but it didn't solve the problem as a debugging console.log under the for loop shows that the function continues to be called multiple times. The callstack on chrome when debugging the button doesn't show any duplicate calls except the one.
I am leaving the bug for now, hopefully I will figure it out some time in the nearest future.